improve: 优化extractor和website设计
