+++ src/alist_mikananirss/core/rss_monitor.py
@@ -3,7 +3,6 @@ import asyncio
 from loguru import logger
 
 from alist_mikananirss.common.database import SubscribeDatabase
-from alist_mikananirss.extractor import Extractor
 from alist_mikananirss.websites import FeedEntry, ResourceInfo, WebsiteFactory
 
 from .download_manager import DownloadManager
@@ -15,7 +14,7 @@ class RssMonitor:
         self,
         subscribe_urls: list[str] | str,
         filter: RegexFilter,
-        extractor: Extractor = None,
+        use_extractor: bool = False,
     ) -> None:
         """The rss feed manager"""
         if not isinstance(subscribe_urls, list):
@@ -25,38 +24,46 @@ class RssMonitor:
             WebsiteFactory.get_website_parser(url) for url in subscribe_urls
         ]
         self.filter = filter
-        self.extractor = extractor
+        self.use_extractor = use_extractor
         self.db = SubscribeDatabase()
         self.interval_time = 300
 
     def set_interval_time(self, interval_time: int):
         self.interval_time = interval_time
 
-    async def get_new_resource(self, fileter: RegexFilter):
+    async def get_new_resources(self, fileter: RegexFilter) -> list[ResourceInfo]:
         new_resources: list[ResourceInfo] = []
         for website in self.websites:
-            feed_entries = await website.get_feed_entries()
+            try:
+                feed_entries = await website.get_feed_entries()
+            except Exception as e:
+                logger.error(f"Failed to get feed entries from {website.rss_url}: {e}")
+                continue
             feed_entries_filted: list[FeedEntry] = []
             for entry in feed_entries:
                 flag = fileter.filt_single(entry.resource_title)
                 if flag:
                     feed_entries_filted.append(entry)
-
             for entry in feed_entries_filted:
                 if not self.db.is_resource_title_exist(entry.resource_title):
-                    resourcec = await website.extract_resource_info(entry)
-                    new_resources.append(resourcec)
+                    try:
+                        resource_info = await website.extract_resource_info(
+                            entry, self.use_extractor
+                        )
+                        new_resources.append(resource_info)
+                    except Exception as e:
+                        logger.error(
+                            f"Pass {entry.resource_title} because of error: {e}"
+                        )
+                        continue
         return new_resources
 
     async def run(self):
         while 1:
             logger.info("Start update checking")
-            new_resources = await self.get_new_resource(fileter=self.filter)
+            new_resources = await self.get_new_resources(fileter=self.filter)
             if not new_resources:
                 logger.info("No new resources")
             else:
-                for resource in new_resources:
-                    if self.extractor:
-                        await self.extractor.process(resource)
                 await DownloadManager.add_download_tasks(new_resources)
             await asyncio.sleep(self.interval_time)
