+++ src/alist_mikananirss/extractor/__init__.py
@@ -1,2 +1,5 @@
-from .extractor import ChatGPTExtractor, Extractor, RegexExtractor
+from .base import ExtractorBase
+from .chatgpt import ChatGPTExtractor
+from .extractor import Extractor
 from .models import AnimeNameInfo, ResourceNameInfo
+from .regex import RegexExtractor
